Fix build warning about size integer comparison.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@1807 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/libsrc/wspiapi/WspiapiLegacyGetNameInfo.c
@@ -21,11 +21,11 @@ WspiapiLegacyGetNameInfo (const struct sockaddr *ptSocketAddress,
   struct in_addr l_inaddr;
   char *pnode = NULL, *pc = NULL;
 
-  if (!ptSocketAddress || tSocketLength < sizeof (struct sockaddr))
+  if (!ptSocketAddress || tSocketLength < (int) sizeof (struct sockaddr))
     return EAI_FAIL;
   if (ptSocketAddress->sa_family != AF_INET)
     return EAI_FAMILY;
-  if (tSocketLength < sizeof (struct sockaddr_in))
+  if (tSocketLength < (int) sizeof (struct sockaddr_in))
     return EAI_FAIL;
   if (!(pszNodeName && tNodeLength) && !(pszServiceName && tServiceLength))
     return EAI_NONAME;
